+++ payment/bank_transfer_test.go
@@ -69,30 +69,3 @@ func TestExecGANBTran(t *testing.T) {
 	result, _ := cli.ExecTranGANB(req)
 	assert.Equal(t, expected, result)
 }
-
-func TestSearchTradeMulti(t *testing.T) {
-	expected := &SearchTradeMultiResponse{
-		Status: TradeMultiStatusPaysuccess,
-	}
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		form := url.Values{}
-		_ = parser.Encoder().Encode(expected, form)
-		w.Header().Set("Content-Type", "application/x-www-form-urlencoded")
-		w.Write([]byte(form.Encode()))
-	}))
-	defer ts.Close()
-	defaultProxy := http.DefaultTransport.(*http.Transport).Proxy
-	http.DefaultTransport.(*http.Transport).Proxy = func(req *http.Request) (*url.URL, error) {
-		return url.Parse(ts.URL)
-	}
-	defer func() { http.DefaultTransport.(*http.Transport).Proxy = defaultProxy }()
-
-	cli, _ := NewClient("siteID", "sitePass", "shopID", "shopPass", false)
-	cli.APIHost = apiHostTest
-	req := &SearchTradeMultiRequest{
-		OrderID: "orderID",
-		PayType: "36",
-	}
-	result, _ := cli.SearchTradeMulti(req)
-	assert.Equal(t, expected, result)
-}
