+++ payment/multi_payment_test.go
@@ -1,0 +1,38 @@
+package payment
+
+import (
+	"net/http"
+	"net/http/httptest"
+	"net/url"
+	"testing"
+
+	"github.com/abyssparanoia/go-gmo/internal/pkg/parser"
+	"gopkg.in/go-playground/assert.v1"
+)
+
+func TestSearchTradeMulti(t *testing.T) {
+	expected := &SearchTradeMultiResponse{
+		Status: TradeMultiStatusPaysuccess,
+	}
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		form := url.Values{}
+		_ = parser.Encoder().Encode(expected, form)
+		w.Header().Set("Content-Type", "application/x-www-form-urlencoded")
+		w.Write([]byte(form.Encode()))
+	}))
+	defer ts.Close()
+	defaultProxy := http.DefaultTransport.(*http.Transport).Proxy
+	http.DefaultTransport.(*http.Transport).Proxy = func(req *http.Request) (*url.URL, error) {
+		return url.Parse(ts.URL)
+	}
+	defer func() { http.DefaultTransport.(*http.Transport).Proxy = defaultProxy }()
+
+	cli, _ := NewClient("siteID", "sitePass", "shopID", "shopPass", false)
+	cli.APIHost = apiHostTest
+	req := &SearchTradeMultiRequest{
+		OrderID: "orderID",
+		PayType: PayEasyPayType,
+	}
+	result, _ := cli.SearchTradeMulti(req)
+	assert.Equal(t, expected, result)
+}
